async_wrap: use double, not int64_t, for uid
The array math in JS converting the in64_t to a safe integer is
unnecessary. Instead set the field to a double and allow internals
direct access to the array to reduce overhead.

+++ src/async-wrap-inl.h
@@ -96,7 +96,7 @@ inline AsyncWrap::ProviderType AsyncWrap::provider_type() const {
 }
 
 
-inline int64_t AsyncWrap::get_uid() const {
+inline double AsyncWrap::get_uid() const {
   return uid_;
 }
